--- token/exporter.py
@@ -1,3 +1,0 @@
-# pylint: disable=unused-import
-from token.token import Token
-from token.token_type_enum import TokenType
