+++ token_type/token.py
@@ -1,5 +1,5 @@
 from typing import Any
-from token.token_type_enum import TokenType
+from token_type.token_type_enum import TokenType
 
 
 class Token:
